Move ingest_array to _dispatcher.py to avoid circular imports and start on BIOM compatibility
