+++ skbio/_dispatcher.py
@@ -1,6 +1,7 @@
 import numpy as np
 import pandas as pd
 
+from skbio.table import Table
 from ._config import get_option
 from ._optionals import _get_polars
 
@@ -35,6 +36,8 @@ def create_table(data, columns=None, index=None, backend=None):
     elif backend == "polars":
         polars = _get_polars()
         return polars.DataFrame(data, schema=columns)
+    # elif backend == "biom":
+    #     return Table(data, observation_ids=index, sample_ids=columns)
     else:
         raise ValueError(f"Unsupported backend: '{backend}'")
 
@@ -62,7 +65,7 @@ def create_table_1d(data, index=None, backend=None):
     if backend is None:
         backend = get_option("tabular_backend")
 
-    if backend == "pandas":
+    if backend in ("pandas"):  # , "biom"):
         return pd.Series(data, index=index)
     elif backend == "numpy":
         return np.array(data)
@@ -71,3 +74,44 @@ def create_table_1d(data, index=None, backend=None):
         return polars.Series(values=data)
     else:
         raise ValueError(f"Unsupported backend: '{backend}'")
+
+
+def ingest_array(input_data, row_ids=None, col_ids=None, dtype=None):
+    # pd.DataFrame
+    if isinstance(input_data, pd.DataFrame):
+        data_ = input_data.values
+        row_ids = list(input_data.index) if row_ids is None else row_ids
+        col_ids = list(input_data.columns) if col_ids is None else col_ids
+    # numpy array
+    elif isinstance(input_data, np.ndarray):
+        data_ = input_data
+    # BIOM (skbio) Table
+    elif isinstance(input_data, Table):
+        # BIOM puts samples as columns and features as rows, so need to handle
+        # accordingly
+        # Or, maybe it's better just to enforce that rows must be samples and
+        # columns must be features (observations), then we don't worry about it
+        print(
+            "BIOM format uses samples as columns and features as rows. Most "
+            "scikit-bio functions expect samples as rows and features as columns."
+            "Please ensure your input is in the correct orientation.\n"
+        )
+        data_ = input_data.to_dataframe().values
+        row_ids = (
+            list(input_data.ids(axis="observation")) if row_ids is None else row_ids
+        )
+        col_ids = list(input_data.ids()) if col_ids is None else col_ids
+    # pl.DataFrame
+    elif hasattr(input_data, "schema"):
+        pl = _get_polars()
+        if isinstance(input_data, pl.DataFrame):
+            data_ = input_data.to_numpy()
+            col_ids = list(input_data.schema) if col_ids is None else col_ids
+    # ndarray
+
+    else:
+        raise TypeError(
+            "Input data must be pandas DataFrame, polars DataFrame, or numpy ndarray"
+        )
+
+    return data_, row_ids, col_ids
