+++ skbio/stats/ordination/_correspondence_analysis.py
@@ -7,13 +7,12 @@
 # ----------------------------------------------------------------------------
 
 import numpy as np
-import pandas as pd
+import pandas as pds
 from scipy.linalg import svd
 
 from ._ordination_results import OrdinationResults
 from ._utils import svd_rank
-from skbio._dispatcher import create_table, create_table_1d
-from skbio.util._misc import ingest_array
+from skbio._dispatcher import create_table, create_table_1d, ingest_array
 
 
 def ca(X, scaling=1, sample_ids=None, feature_ids=None, output_format=None):
